requirements.txt
pytorch
smdebug.pytorch
torch
torchvision
sys
pip